+++ index.html
@@ -655,6 +655,6 @@
       validateForm3.init();
     </script> -->
 
-    <script src="./scripts/script.js"></script>
+    <script src="./dist/main.js"></script>
   </body>
 </html>
