+++ src/modules/checkCalcInputs.js
@@ -1,0 +1,11 @@
+function checkCalcInputs() {
+  document.querySelectorAll('.calc-block input').forEach(item => {
+    item.addEventListener('input', event => {
+      const input = event.target;
+
+      input.value = input.value.replace(/[^\d]/, '');
+    });
+  });
+}
+
+export default checkCalcInputs;
